Move/rename creator for startPeriod action
