+++ src/slices/live/live-action-creators.ts
@@ -2,7 +2,13 @@
 
 import { Duration } from '../../models/clock.js';
 import { FormationBuilder, getPositions } from '../../models/formation.js';
-import { LiveGame, LivePlayer, PeriodStatus, getPlayer } from '../../models/live.js';
+import {
+  LiveGame,
+  LivePlayer,
+  PeriodStatus,
+  gameCanStartPeriod,
+  getPlayer,
+} from '../../models/live.js';
 import { PlayerStatus } from '../../models/player.js';
 import { ThunkResult } from '../../store.js';
 import { isPeriodOverdue } from './clock-reducer-logic.js';
@@ -16,7 +22,24 @@ const {
   invalidStarters,
   markPeriodOverdue,
   startersCompleted,
+  startPeriod,
 } = actions;
+
+export const startPeriodCreator =
+  (gameId: string): ThunkResult =>
+  (dispatch, getState) => {
+    const state = getState();
+    const game = selectLiveGameById(state, gameId);
+    if (!game || !game.clock) {
+      return;
+    }
+    dispatch(
+      startPeriod(
+        game.id,
+        gameCanStartPeriod(game, game.clock.currentPeriod, game.clock.totalPeriods)
+      )
+    );
+  };
 
 export const markPeriodOverdueCreator =
   (gameId: string): ThunkResult =>
